README lint cleanup
Unintentional double spaces removed: 1
All extra spacing in front of lines are required for the information to appear as intended. I believe no other changes can be made without compromising the formatting. If you know of a different way to achieve the same or similar results with less leading spaces in markdown please let me know.