GH BT scanning is wonky, you do it in 2 steps telling it to scan and then checking the results.
The timeout on the start scan command is actually how long to scan for, so really longer is actually better. But I don't know if there is issues with having it run forever.

The other problem is that you either have it set to clear_results or not, if false it never clears out devices that are no longer visible which doesn't work cause we can't tell if a device is gone. But if its true a device if it misses one scan it will go unavailable right away and might not show up a while later. So to solve this I do a clear_results every minute. But I am also using last_seen to not set a device as gone until that point. I could probably just do the last_seen, but I think both together seems safer at least to start.